refactor connectionToEdgeId to utils file
