+++ app/lib/node-editor/node-store/utils.ts
@@ -1,3 +1,5 @@
+import type { Connection } from "@xyflow/react";
+
 import type { nodeInputs } from "./node-store";
 
 export const getInput = (
@@ -8,3 +10,14 @@ export const getInput = (
   const input = inputs.get(handleId);
   return input?.sourceNode.getResult(input.sourceHandleId) ?? fallback;
 };
+
+export function connectionToEdgeId(edge: Connection): string {
+  return (
+    "xy-edge__" +
+    edge.source +
+    edge.sourceHandle +
+    "-" +
+    edge.target +
+    edge.targetHandle
+  );
+}
